Renamed: component ShowWatchedProducts.vue
