+++ apps/frontend/src/components/page-products/ShowWatchedProducts.vue
